pyproject.toml
[project]
name = "reversi"
version = "0.1.0"
description = "A simple reversi game in pygame"
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pygame-ce>=2.5.3"]

[project.urls]
Repository = "https://github.com/blazingyeti/reversi"

[project.optional-dependencies]
dev = ["ruff", "mypy", "build", "pyinstaller"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
reversi = "src.main:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.ruff]
target-version = "py313"
indent-width = 4
line-length = 100
exclude = ["*.json"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "C90"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[tool.mypy]
disallow_untyped_defs = true
python_version = "3.13"
warn_return_any = true
strict = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pyright]
venvPath = "."
venv = ".venv"
